requirements.txt
asgiref==3.4.1
boto3==1.17.23
botocore==1.20.23
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==2.2.12
django-cors-headers==3.2.0
django-crispy-forms==1.11.1
django-haystack==3.1.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-machina==1.1.5
django-mptt==0.13.4
django-newsletter==0.9.1
django-redis-cache==2.1.0
django-rq==2.1.0
django-storages==1.11.1
django-tinymce==3.3.0
django-translation-manager==1.1.1
django-widget-tweaks==1.4.9
djangorestframework==3.9.4
facebookads==2.11.4
gunicorn==20.0.4
html5lib==1.1
idna==3.3
jmespath==0.10.0
ldif3==3.1.1
markdown2==2.4.1
oauthlib==3.1.1
Pillow==8.1.2
polib==1.1.1
psycopg2==2.9.2
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.3.0
PyPDF2==1.26.0
python-card-me==0.9.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
reportlab==3.6.3
requests==2.26.0
requests-oauthlib==1.3.0
rq==1.10.0
s3transfer==0.3.4
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
surlex==0.2.0
unicodecsv==0.14.1
urllib3==1.26.3
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.3
